# Speedup MLARopeQuantize by 20-40%